Add files via upload
This code creates a directed acyclic graph with a fixed set of vertices and edges. It then attempts to generate a random topological sorting of the vertices in the graph using a loop that runs 3 times. If the graph contains no cycle, the sorted list of vertices is printed. If the graph contains a cycle, a message indicating this is printed. The random module is used to randomly shuffle vertices and break ties in the sorting process. Finally, the graph can be displayed using the commented out lines of code.